// apps/deckgl-react/components/control-panel.module.css.d.ts
declare const styles:
  & Readonly<{ "container": string }>
  & Readonly<{ "dropDownContainer": string }>
  & Readonly<{ "dropDownLabel": string }>
  & Readonly<{ "dropDown": string }>
  & Readonly<{ "tilesetDropDownContainer": string }>
  & Readonly<{ "tilesetDropDownLabel": string }>
  & Readonly<{ "tilesetDropDown": string }>
  & Readonly<{ "checkBoxContainer": string }>
  & Readonly<{ "checkBoxLabel": string }>
  & Readonly<{ "checkBox": string }>
;
export default styles;
//# sourceMappingURL=./control-panel.module.css.d.ts.map